pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-agent-ecommerce"
version = "1.0.0"
description = "AI Multi-Agent System for Warehouse and Marketplace Integration"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Multi-Agent Team", email = "team@multiagent.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Financial :: Point-Of-Sale",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
keywords = ["ecommerce", "multi-agent", "ai", "warehouse", "marketplace"]
requires-python = ">=3.9"

dependencies = [
    # Core dependencies
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    
    # Messaging
    "aiokafka>=0.10.0",
    "kafka-python>=2.0.0",
    
    # Cache
    "redis>=5.0.0",
    "aioredis>=2.0.0",
    
    # HTTP clients
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "requests>=2.31.0",
    
    # Security
    "cryptography>=41.0.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "PyJWT>=2.8.0",
    
    # AI/ML
    "openai>=1.3.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "pandas>=2.1.0",
    
    # Monitoring
    "prometheus-client>=0.19.0",
    "structlog>=23.2.0",
    
    # CLI
    "click>=8.1.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    
    # Configuration
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    
    # Utilities
    "python-multipart>=0.0.6",
    "email-validator>=2.1.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
]

test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.25.0",
    "faker>=20.1.0",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/your-org/multi-agent-ecommerce"
Documentation = "https://multi-agent-ecommerce.readthedocs.io"
Repository = "https://github.com/your-org/multi-agent-ecommerce.git"
Issues = "https://github.com/your-org/multi-agent-ecommerce/issues"

[project.scripts]
multi-agent = "multi_agent_ecommerce.cli:main"
multi-agent-start = "multi_agent_ecommerce.agents.start_agents:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["multi_agent_ecommerce*", "agents*", "shared*", "database*"]

[tool.setuptools.package-data]
"*" = ["*.yml", "*.yaml", "*.json", "*.toml", "*.txt", "*.md"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["multi_agent_ecommerce", "agents", "shared", "database"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "e2e: marks tests as end-to-end tests",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["multi_agent_ecommerce", "agents", "shared", "database"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.mypy]
python_version = "3.9"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "kafka.*",
    "aiokafka.*",
    "redis.*",
    "aioredis.*",
]
ignore_missing_imports = true
